+++ src/main/kotlin/com/github/olivernybroe/wingidea/ide/highlighting/WingSyntaxHighlighter.kt
@@ -27,6 +27,7 @@ class WingSyntaxHighlighter: SyntaxHighlighterBase() {
             put(WingElementTypes.MULTI_LINE_COMMENT, WingColors.BLOCK_COMMENT)
             put(WingElementTypes.SEMICOLON, WingColors.SEMI_COLON)
 
+
             fillMap(this, WING_KEYWORDS, WingColors.KEYWORD)
             fillMap(this, WING_BRACES, WingColors.BRACES)
             fillMap(this, WING_BRACKETS, WingColors.BRACKETS)